+++ tests/module_disambiguation/expected_both/syntax.expanded.rs
@@ -1,3 +1,4 @@
+use duplicate::*;
 mod module_some_name11 {
     pub struct SomeName11();
 }