+++ tests/module_disambiguation/from/verbose_syntax.rs
@@ -1,5 +1,6 @@
+use duplicate::*;
 // Tests module names are postfixed from substitution identifier
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		name [SomeName11]
 	]
@@ -17,7 +18,7 @@ mod module {
 
 // Tests if multiple identifiers are given, the first identifier who's substitutions
 // all are simple identifiers (and nothing else) is chosen
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		member_type	[Vec<()>]
 		name 		[SomeName21]
@@ -38,7 +39,7 @@ mod module {
 
 // Tests if multiple identifiers are given, the first identifier who's substitutions
 // all are simple identifiers is chosen
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		member_type		[u8]
 		filler_ident	[SomeIdent]
@@ -61,7 +62,7 @@ mod module {
 //item_end
 
 // Tests only the module name is disambiguated and not any identifiers used inside it.
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		name [SomeName41]
 	]