+++ tests/no_features/expected/insert.expanded.rs
@@ -1,3 +1,4 @@
+use duplicate::*;
 pub struct SomeName();
 pub struct SomeName2(i32);
 pub struct SomeName3(&i32);