+++ tests/no_features/expected_both/from_macro.expanded.rs
@@ -1,3 +1,4 @@
+use duplicate::*;
 pub struct SomeName1();
 pub struct SomeName2(SomeMember2);
 pub struct SomeName3(SomeMember3);