+++ tests/no_features/expected_both/nested_duplicate.expanded.rs
@@ -1,3 +1,4 @@
+use duplicate::*;
 pub struct SomeName1();
 pub struct SomeName2();
 pub struct SomeName3();