+++ tests/no_features/from/nested_duplicate.rs
@@ -1,5 +1,6 @@
+use duplicate::*;
 // Tests nesting in substitutions
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	members [
 		duplicate!{
 			[
@@ -20,7 +21,7 @@ impl Example {
 //item_end
 
 // Tests nesting between global substitutions
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	name [StructName1];
 	duplicate!{
 		[
@@ -34,7 +35,7 @@ struct name(typ1,typ2);
 //item_end
 
 // Tests nesting between individual substitutions
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	name typ1 typ2;
 	[TypeName21]
 	duplicate!{
@@ -46,4 +47,30 @@ struct name(typ1,typ2);
 	[TypeName23];
 )]//duplicate_end
 struct name(typ1, typ2);
+//item_end
+
+// Tests sequential nesting
+#[duplicate_item(
+	v1;	[31]; [32];
+)]//duplicate_end
+#[duplicate_item(
+	v2;	[33, v1]; [34, v1];
+)]//duplicate_end
+impl SomeType<v2> for (){}
+//item_end
+//item_end
+
+// Tests sequential nesting (3-deep
+#[duplicate_item(
+	v1;	[41]; [42];
+)]//duplicate_end
+#[duplicate_item(
+	v2;	[43, v1]; [44, v1];
+)]//duplicate_end
+#[duplicate_item(
+	v3;	[45, v2]; [46, v2];
+)]//duplicate_end
+impl SomeType<v3> for (){}
+//item_end
+//item_end
 //item_end