+++ tests/no_features/from/short_from_macro.rs
@@ -4,10 +4,11 @@
 // Each test consists of a macro_rules declaration which uses
 // some specific macro variable type to no_features to the duplicate invocation.
 // Then the created macro is invoked.
+use duplicate::*;
 
 macro_rules! test_ident_from_macro_variable{
 	{ $name:ident } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$name;	[SomeName1]
 		)]//duplicate_end
 		pub struct $name();
@@ -18,7 +19,7 @@ test_ident_from_macro_variable!(name);
 
 macro_rules! test_2_idents_from_macro_variable{
 	{ $($idents:ident)* } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$($idents)*;
 			[SomeName2] [SomeMember2]
 		)]//duplicate_end
@@ -30,7 +31,7 @@ test_2_idents_from_macro_variable!(name member);
 
 macro_rules! test_ident_from_macro_path_variable{
 	{ $name:path } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$name;	[SomeMember3]
 		)]//duplicate_end
 		pub struct SomeName3($name);
@@ -41,7 +42,7 @@ test_ident_from_macro_path_variable!(name);
 
 macro_rules! test_ident_from_macro_expr_variable{
 	{ $name:expr } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$name;	[SomeValue4]
 		)]//duplicate_end
 		const SomeName4: () = $name;
@@ -52,7 +53,7 @@ test_ident_from_macro_expr_variable!(name);
 
 macro_rules! test_ident_from_macro_type_variable{
 	{ $name:ty } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$name;	[SomeMember5]
 		)]//duplicate_end
 		pub struct SomeName5($name);
@@ -63,7 +64,7 @@ test_ident_from_macro_type_variable!(name);
 
 macro_rules! test_ident_from_macro_pattern_variable{
 	{ $name:pat } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$name;	[SomeName6]
 		)]//duplicate_end
 		fn some_fn6(){
@@ -76,7 +77,7 @@ test_ident_from_macro_pattern_variable!(name);
 
 macro_rules! test_ident_from_macro_statement_variable{
 	{ $name:stmt } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$name;	[SomeName7]
 		)]//duplicate_end
 		fn some_fn7(){
@@ -89,7 +90,7 @@ test_ident_from_macro_statement_variable!(name);
 
 macro_rules! test_ident_from_macro_token_tree_variable{
 	{ $name:tt } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			$name;	[SomeName8]
 		)]//duplicate_end
 		pub struct $name();