+++ tests/no_features/from/verbose_from_macro.rs
@@ -4,10 +4,11 @@
 // Each test consists of a macro_rules declaration which uses
 // some specific macro variable type to no_features to the duplicate invocation.
 // Then the created macro is invoked.
+use duplicate::*;
 
 macro_rules! test_ident_from_macro_variable{
 	{ $name:ident } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$name	[SomeName1]
 			]
@@ -20,7 +21,7 @@ test_ident_from_macro_variable!(name);
 
 macro_rules! test_2_idents_from_macro_variable{
 	{ $($idents:ident)*,  $($tts:tt)*} => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$($idents[$tts])*
 			]
@@ -33,7 +34,7 @@ test_2_idents_from_macro_variable!(name member, SomeName2 SomeMember2);
 
 macro_rules! test_ident_from_macro_path_variable{
 	{ $name:path } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$name	[SomeMember3]
 			]
@@ -46,7 +47,7 @@ test_ident_from_macro_path_variable!(name);
 
 macro_rules! test_ident_from_macro_expr_variable{
 	{ $name:expr } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$name	[SomeValue4]
 			]
@@ -59,7 +60,7 @@ test_ident_from_macro_expr_variable!(name);
 
 macro_rules! test_ident_from_macro_type_variable{
 	{ $name:ty } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$name	[SomeMember5]
 			]
@@ -72,7 +73,7 @@ test_ident_from_macro_type_variable!(name);
 
 macro_rules! test_ident_from_macro_pattern_variable{
 	{ $name:pat } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$name	[SomeName6]
 			]
@@ -87,7 +88,7 @@ test_ident_from_macro_pattern_variable!(name);
 
 macro_rules! test_ident_from_macro_statement_variable{
 	{ $name:stmt } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$name	[SomeName7]
 			]
@@ -102,7 +103,7 @@ test_ident_from_macro_statement_variable!(name);
 
 macro_rules! test_ident_from_macro_token_tree_variable{
 	{ $name:tt } => {
-		#[duplicate::duplicate_item(
+		#[duplicate_item(
 			[
 				$name	[SomeName8]
 			]