+++ tests/no_features/from/verbose_ident_argument.rs
@@ -1,5 +1,6 @@
+use duplicate::*;
 // Test single-token argument
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name		[fn_const_1]
 		refs(type)	[&type]
@@ -13,7 +14,7 @@ fn fn_name(arg: refs([i32])){}
 //item_end
 
 // Test multi-token argument
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name		[fn_const_2]
 		refs(type)	[&type]
@@ -27,7 +28,7 @@ fn fn_name<'a>(arg: refs([Vec<i32>])){}
 //item_end
 
 // Test multi-argument identifier
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name					[fn_const_3]
 		refs(lifetime, type)	[& 'lifetime type]
@@ -41,7 +42,7 @@ fn fn_name<'a>(arg: refs([a],[i32])){}
 //item_end
 
 // Test multi-argument identifier and multi-token arguments
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name					[fn_const_4]
 		refs(lifetime, type)	[& 'lifetime type]
@@ -55,7 +56,7 @@ fn fn_name<'a>(arg: refs([a],[Result<i32,u8>],)){}
 //item_end
 
 // Test multiple invocations of identifiers with arguments
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name					[fn_const_5]
 		refs(lifetime, type)	[& 'lifetime type]
@@ -69,7 +70,7 @@ fn fn_name<'a>(arg: refs([a], [Result<i32,u8>],))  -> refs([a], [i32]) {}
 //item_end
 
 // Test the same identifier argument can have different names in different substitution groups.
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name			[fn_const_6]
 		refs(type_1)	[&type_1]
@@ -84,7 +85,7 @@ fn fn_name(arg: refs([i32])){}
 
 // Test that identifiers with arguments don't have to come in the same order in different
 // substitution groups.
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name		[fn_const_7]
 		refs(type)	[&type]
@@ -98,7 +99,7 @@ fn fn_name(arg: refs([i32])){}
 //item_end
 
 // Test multiple identifiers with arguments
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name		[fn_const_8]
 		refs(type)	[&type]
@@ -114,7 +115,7 @@ fn fn_name(arg: refs([i32]), second_arg: arg_2([i64])){}
 //item_end
 
 // Test identifier with argument called inside itself
-#[duplicate::duplicate_item(
+#[duplicate_item(
 	[
 		fn_name		[fn_const_9]
 		refs(type)	[&type]