+++ tests/utils.rs
@@ -112,8 +112,9 @@ impl<'a> ExpansionTester<'a>
 		}
 
 		// Prepare feature list for expansion testing
-		let mut args: Vec<&str> = Vec::new();
+		let mut args = Vec::new();
 		let mut features = String::new();
+		args.push("--no-default-features");
 		if NR_FEATURES > 0
 		{
 			args.push("--features");
@@ -152,8 +153,8 @@ impl<'a> ExpansionTester<'a>
 	/// Generates an action that creates two versions of the given file in the
 	/// testing directory. The source file must use the 'duplicate' attribute
 	/// macro, where:
-	/// - The invocation must starts with `#[duplicate::duplicate_item(` on a
-	///   the first line
+	/// - The invocation must starts with `#[duplicate_item(` on a the first
+	///   line
 	/// (with nothing else). Notice that you must not import the attribute but
 	/// use its full path.
 	/// - Then the body of the invocation. Both syntaxes are allowed.
@@ -175,7 +176,7 @@ impl<'a> ExpansionTester<'a>
 	/// ### Example
 	/// Original file (`test.rs`):
 	/// ```
-	/// #[duplicate::duplicate_item(
+	/// #[duplicate_item(
 	///   name;
 	///   [SomeName];
 	/// )]//duplicate_end
@@ -184,7 +185,7 @@ impl<'a> ExpansionTester<'a>
 	/// ```
 	/// First version (`test.expanded.rs`):
 	/// ```
-	/// #[duplicate::duplicate_item(
+	/// #[duplicate_item(
 	///   name;
 	///   [SomeName];
 	/// )]
@@ -192,7 +193,7 @@ impl<'a> ExpansionTester<'a>
 	/// ```
 	/// Second version (`inline_test.expanded.rs`):
 	/// ```
-	/// duplicate::duplicate{
+	/// duplicate{
 	///   [
 	///     name;
 	///     [SomeName];
@@ -222,13 +223,11 @@ impl<'a> ExpansionTester<'a>
 
 				match line
 				{
-					"#[duplicate::duplicate_item(" =>
+					"#[duplicate_item(" =>
 					{
-						dest_file
-							.write_all("#[duplicate::duplicate_item(".as_bytes())
-							.unwrap();
+						dest_file.write_all("#[duplicate_item(".as_bytes()).unwrap();
 						dest_inline_file
-							.write_all("duplicate::duplicate!{\n[".as_bytes())
+							.write_all("duplicate!{\n[".as_bytes())
 							.unwrap();
 					},
 					")]//duplicate_end" =>