+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `duplicate` now also substitutes the invocations of nested `duplicate`'s. See [#48](https://github.com/Emoun/duplicate/issues/48).
+
 ## [0.4.1] - 2022-07-17
 
 ### Fixed